Merge in TAB/tabula from task/es-7-tests to develop

* commit '5c689bb6f5707d8729fdb8c15c98fcb9b454e476':
  Upgrade tests to use Elasticsearch 7 (with TestContainers instead of unsupported elastic4s-embedded). Upgrade dependencies

+++ common/src/main/scala/uk/ac/warwick/tabula/services/healthchecks/ImportStatusHealthchecks.scala
@@ -84,7 +84,7 @@ abstract class AbstractImportStatusHealthcheck(name: String)
       LocalDateTime.now(DateTimeUtils.CLOCK_IMPLEMENTATION),
       message,
       Seq[ServiceHealthcheck.PerformanceData[_]](
-        new ServiceHealthcheck.PerformanceData("last_successful_hours", lastSuccessfulHoursAgo, WarningThreshold.toHours, ErrorThreshold.toHours)
+        new ServiceHealthcheck.PerformanceData("last_successful_hours", lastSuccessfulHoursAgo, WarningThreshold.toHours.toDouble, ErrorThreshold.toHours.toDouble)
       ).asJava
     )
   }
